Move navigator component
