+++ app/navigator/props.ts
@@ -1,0 +1,3 @@
+export default interface Props {
+  bgColor: string;
+}
