+++ app/navigator/style.ts
